util: move tensor utilities to tensor_util.ts
