+++ src/disk_experiment_test.ts
@@ -18,7 +18,7 @@ import * as path from "path";
 import * as rimraf from "rimraf";
 import { test } from "../tools/tester";
 import { DiskExperiment, isDir } from "./disk_experiment";
-import { assert, assertAllEqual } from "./util";
+import { assert, assertAllEqual } from "./tensor_util";
 
 function setup() {
   process.env.PROPEL_DIR = path.join(os.tmpdir(), "propel_test");
