+++ src/npy_test.ts
@@ -1,8 +1,9 @@
 import { test } from "../tools/tester";
 import * as pr from "./api";
 import * as npy from "./npy";
+import * as util from "./tensor_util";
 import * as types from "./types";
-import * as util from "./util";
+import { IS_NODE } from "./util";
 
 // Because the python interop test requires python and numpy installed
 // it is turned off by default here. The option is left for debugging
@@ -47,7 +48,7 @@ test(async function npy_serialize() {
   util.assertAllEqual(t, tt);
 });
 
-if (util.IS_NODE && testPython) {
+if (IS_NODE && testPython) {
   test(async function npy_pythonInterop() {
     await checkPython("[ 1.5  2.5]", [ 1.5, 2.5 ]);
     await checkPython("[ 1.  2.]", [ 1, 2 ]);
