+++ src/ops.ts
@@ -19,8 +19,8 @@
 import { bo } from "./backend";
 import * as backprop from "./backprop";
 import { convert, Tensor } from "./tensor";
+import { assert, bcastGradientArgs, shapesEqual } from "./tensor_util";
 import * as types from "./types";
-import { assert, bcastGradientArgs, shapesEqual } from "./util";
 
 // FWFunc defines a "primative" op (using autograd nomenclature). It should
 // never use Tensors, only BasicTensors. These forward pass functions
