+++ tools/test_isomorphic.ts
@@ -7,4 +7,5 @@ import "../src/mnist_test";
 import "../src/nn_example_test";
 import "../src/npy_test";
 import "../src/params_test";
+import "../src/tensor_util_test";
 import "../src/util_test";
