+++ tools/test_node.ts
@@ -1,5 +1,6 @@
-import "../src/disk_experiment_test";
 import "./test_isomorphic";
+
+import "../src/disk_experiment_test";
 
 // Only on Node/TF should we run the binding_test.
 import { backend } from "../src/api";
